html
Not much to be said here we are setting some basic html headers. The Javascript will create the content that goes into the div with the id of logger.

The dice img tags have data-dice on them that help the javascript to know exactly what number they need to roll without doing a find replace on the id. This seems like something that could be optimized later.